Move static semantics into separate module
